refactor(#67): type 폴더로 PostList 인터페이스 이동
